requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
duckdb==1.4.3
idna==3.11
numpy==2.4.1
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML>=6.0.1
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.3

# ML dependencies (TKT-137, TKT-142)
scikit-learn>=1.4.0
scipy>=1.12.0

# Data validation (IDEA-006)
great_expectations>=1.0.0

# Development dependencies
ruff>=0.8.0
pytest>=8.0.0
pytest-cov>=5.0.0
pre-commit>=4.0.0
mypy>=1.11.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Claude Orchestrator (multi-agent API)
anthropic>=0.49.0
python-dotenv>=1.0.0

# ML Explainability - optional, requires numba/llvmlite compilation
# Install separately: pip install shap>=0.42.0
# shap>=0.42.0
